Rename file without leading underscore
the folder it's in is already private
