+++ src/telegram/_files/inputstorycontent.py
